+++ src/extension.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode';
+import { formatPendingChanges } from './commands/formatPendingChanges';
 import { ExtensionContext } from './extensionContext';
 import { registerDocumentFormattingEditProvider } from './providers/documentFormattingProvider';
 import { registerRangeFormattingEditProvider } from './providers/rangeFormattingEditProvider';
@@ -6,6 +7,13 @@ import { registerRangeFormattingEditProvider } from './providers/rangeFormatting
 export function activate(extensionContext: vscode.ExtensionContext) {
   const context = new ExtensionContext(extensionContext);
   context.log.info('Activating extension', extensionContext.extension.packageJSON.version);
+
+  const {
+    commands: { registerCommand: cr },
+  } = vscode;
+  context.disposables.push(
+    cr('formatto.formatPendingChanges', () => formatPendingChanges(context)),
+  );
 
   registerDocumentFormattingEditProvider(context);